Move ProxyComponentResource into Node SDK
